fix DRY
